tdf#130857 qt weld: Set menu button's menu defined in .ui file
Make the existing local helper function `extractPopupMenu`
used by VclBuilder a static method in the BuilderBase base
class for reuse by QtBuilder.

Use it in QtBuilder::makeObject to evaluate
the "popup" property for a "GtkMenuButton"
object, get the corresponding QMenu and set
it for the button.

The code currently assumes that the corresponding
menu is aleady known, i.e. was defined earlier in
the .ui file than the menu button it belongs to.
VclBuilder doesn't rely on that, so a handling
similar to there could be implemented as and when
needed.

Change-Id: I95446fb7a369f2cfee0d8d51432933582bbfd570
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/178972
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>