Re-write ScTable::GetRowForHeight
Cleanup up this method, making it more organised in how it iterates over
spans,
and adding asserts to verify assumptions.
The asserts uncovered that we are sometimes fed bad data, so now we deal
with
that (and not sanitising the input is possibly why the original code was
a little weird).

Also, now that we are returning correct data, our caller does not need
to correct the return value

Change-Id: I094931bb2e93a3273b7a928fc7e83eed75e3db51
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179040
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>