crash reporting: use SfxRequest to get parent for dialogs
0: return pWindow ? pWindow->GetFrameWeld() : nullptr;
2: SfxNewFileDialog aNewFileDlg(GetView()->GetFrameWeld(), SfxNewFileDialogMode::LoadTemplate);

Reproducible by adding "LoadStyles" to the menu in writer, then print
preview and then use that "LoadStyles" entry.

Also dispatch "SID_TEMPLATE_LOAD", as used by "Load Styles from Template"
entry in the dropdown from the "Styles Action" toolbar of the style
panel, with an explicit frame it is associated with.

Change-Id: I3d536b897fbf6be9e20e4238e6a761df56ac9701
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179084
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins