tdf#130857 qt weld: Support "Name Dialog"
Declare support for the name dialog that e.g. gets used
in "Tools" -> "Options" -> "Appearance" when clicking the
"New" button.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Change-Id: Idaa3190d45efba87a74ef80a2999f3c6be7fa90d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179110
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>