tdf#164233 vcl: Fix and simplify spin field mouse over check
Instead of checking individually whether the mouse pos is
over the SpinField's up or down button or the `mpEdit`,
just check the SpinField's whole rectangle.

The mouse event pos is relative to the SpinField, so
use (0, 0) for the top left position of the SpinField's
rect used in the check.

This is not only simpler, but also fixes the case
of the UNO control when the "Spin Button" options is disabled
in which case there are no buttons and `mpEdit` is null, so
`bMouseHovered` would always be false and adjusting the spin field
value for that control using the mouse wheel wouldn't even work when
mouse-hovered after the initial check was introduced in

    commit 869b88488ac443cc64943254064da20b0f361c56
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon May 13 16:37:45 2024 +0200

        tdf#160824 vcl: Require mouse over spinfield to mouse-wheel through values

Change-Id: I3a2ee3af8af872e4a330a4d5e9f5cd6a2de89754
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179085
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>