tdf#130857 qt weld: Introduce QtInstancePopover
When encountering a "GtkPopover" object in a .ui
file, create a QWidget with window flags Qt::Popup
and set a layout, as weld::Popover is a weld::Container
subclass.

Add new QtInstancePopover that QtInstanceBuilder::weld_popover
creates.
For now, methods in that class simply trigger an assert;
actual logic still needs to be implemented.

Change-Id: I9a0dfefd9fb79d24fe40dd5efe13ef0c3ab9b7fb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179098
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>