tdf#132111: don't claim a number format for multiselection / no set format
The "not set" state for RES_BOXATR_FORMAT is handled explicitly in many
places, e.g., in SwTableBox::HasNumContent. This allows to detect number
entered into cells without explicit format. Thus, this state is not the
same as "text" number format, and it must not be selected in the Number
Format dialog. The same is for multiselection with different formats.

Change-Id: I4f4aa66a2cbfd1bf2efa8b99a6e7a2cf9c119712
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179177
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
(cherry picked from commit d74738ef156b6c468fd8aa0fb9e0528f04cf5c07)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179201
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>