pdf: fix PDF encryption and re-enable the test
Slightly wrong loop exit condition when calculating the R6 hash,
which caused that every now and then the document could not be
decrypted correctly. This happened at least 1 time in 100 tries
(usually it did not even take that much to encountered a test
failure), but with the fix it worked correctly for 500 tries, so
it seems this issue is now fixed.

Change-Id: I5fa046892f36dabc367aabed1295802e805ee6e4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179299
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>