Fix menubar back and menu highlight reading black color from the system in GTK
Themeing code gets menubar background color from StyleSettings, which
was reported as rgba[000000ff] by the gtk code. fixed that by adding
a background class to the menubar's GtkStyleContext.

The Highlight color for menubar items and menu items was also reported
rgba[000000ff]. Using Accent color for that for the time being.

Change-Id: I605a75775f2786fe5ee7645ca707f5dd2037040d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179407
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>
Tested-by: Jenkins