tdf#164420 Fix unselected Check/Radio buttons not following themes in GTK
The button part of the check and radio buttons wasn't using theme
colors, because till now I couldn't figure out what css nodes/properties
it was using. Turns out the background for the buttons uses the
background-image property. Set that to none with separator color as the
border color, as that seemed to be the most sensible option.

We might want to theme these too some day.

Change-Id: I90a7fe5965d97e1437e13fe8e2f2a277a5b82485
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179486
Tested-by: Jenkins
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>