Related: tdf#159995 allow setting the page size when printing to file
On macOS, printing to file just saves to PDF using a native print
job just like the native print dialog's PDF > Save as PDF... option.
Also, as noted in tdf#163558, the native print dialog could support
changing the page size eventually (and already does in Apple's
Safari and TextEdit applications) so allow changing of the page
size in the non-native print dialog on macOS as well.

Also, changing a listbox's active item does not trigger an update
to the preview so force an update to the preview by calling each
active listbox's select handler after changing the active printer.

Lastly, when printing to file with the non-native print dialog, try
to mimic the save dialog behavior when using the native print dialog
to save to PDF.

Change-Id: I5f42d818701cf64fdf3ecb616d1eef7b835eb2f8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179573
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>