tdf#164201 docx convert-to: prefer Word 2010–365 Document, not 2007
This was made more prominent by mstahl's 7.6
commit e66ddcd4b66923bc835bd7c5f5c784a809a420a2,
which rightly forced the compatibilityMode to 12 for Word 2007 exports.

The context here is running a command-line
    soffice --convert-to docx myfile.odt
which was creating a compat12 Word 2007 DOCX
instead of a compat15 Word 2013+ DOCX.

Filtercache registers valid conversion types in a seemingly random order
"but if it's a 'Preferred' one - set it to the front of the list.
 Of course, multiple 'Preferred' registrations can occur
 (they shouldn't, but they can!) ..."

For DOCX, both writer_MS_Word_2007_XML and writer_OOXML
say that they are the preferred one,
and since Word 2007 was processed last, it won.

However, since 2020 we have preferred exporting as compat15 (2013+)
so make the modern format the preferred filter.

[I looked for more "multiples", but more often than not,
 NOTHING is set as the preferred entry, so first come first served.]

Change-Id: I2e4c95520619e0aa2e1f0d8ae5e8928d0033902e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179588
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins