tdf#162695 decrease the text margins to smaller
This decreases the text margins so the comment is overall smaller,
which should make the situation better, but the main issue might
be that the comments previously did not change size with zoom, but
now do. This is probably nto so simple to change.

Change-Id: I5c83b787f58324431e110ee095f61135b67588e8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179662
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit e70afe758eae3acdc5ac6ee2b3d173985899db1a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179704
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>