tdf#161480 use BmpScaleFlag::NearestNeighbor when upscaling on macOS
On macOS, due to Retina window scaling, it is very common to
have Skia surfaces that are 2 times the size of their respective
output device sizes so upscaling is often the default state.

But when upscaling bitmaps on macOS with either Skia/Metal or
Skia/Raster, some distortion to color and/or alpha values is
introduced and only BmpScaleFlag::NearestNeighbor will upscale
the bitmap correctly.

Change-Id: I9cf02d3acace82751fa6e114aaecf23cbbb85c65
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179848
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>