tdf#164619: bite the bullet, and write the correct ODF version
The problem was the inconsistency between ODF versions for the forms
in the package and the form's content.xml. The check
for consistency was introduced in version 5.4 in commit
0a0f205775807d63e7f6dd910de368bb12b11a49 (tdf#105686 - fixed wrongly
imported named expressions:, 2017-03-21). In version 7.0, commit
9fdf8c0a5cc036ea9bd1e11dd8f2c1a6e601fae2 (Also consider saved
exceptions when terminating parse, 2019-11-16) made the exception
from that check fatal.

Commit 519d96fd8b83ef4c61576d87b58f97b7e6e6e3c6 (tdf#134582 sfx2:
when storing, set Version on embedded object storage, 2020-07-10)
started to write the embedded object versions to package manifest,
and commit 16de54a5c47fbc4691ee099c1f7bb559a8fe11ac (tdf#138209 ODF
export: work around forms problem in LO < 7.0, 2021-05-10) created
an exception for forms, to workaround the problem in older versions.
But the latter commit created the reverse inconsistency, where the
newer versions started to write older ODF version in manifest.xml,
but newer ODF version in content.xml.

This change restores the correct version written to the manifest.
It is already messy in the already-released versions; it doesn't
look possible to restore sanity retroactively, so just make it
correct from now on, and accept the problem in older versions.

Change-Id: If5d07b6000b9f08b33844db48b815881bed2915b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179867
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins