Move setting of LD_LIBRARY_PATH closer to invocation of cppunittester
At least for a Linux build of mine, `make CppunitTest_vcl_gtk3_a11y` would fail
with something like

> workdir/LinkTarget/Executable/cppunittester X11 error: Can't open display:
>    Set DISPLAY environment variable, use -display option
>    or check permissions of your X-Server
>    (See "man X" resp. "man xhost" for details)

apparently because that test adds an invocation of xvfb-run to ICECREAM_RUN,
which appeared on the command line after gb_CppunitTest_CPPTESTPRECOMMAND
(which, on Linux, sets up LD_LIBRARY_PATH to include instdir/program), so that
the execution of xvfb-run picked up libraries from instdir/program instead of
from the system (at least libpixman-1.so.0, it appears), which then apparently
caused things to not work well.

So move the contents of gb_CppunitTest_CPPTESTPRECOMMAND further down on the
command line (which then requires an invocation of `env`).  (This still leaves
me with a `make CppunitTest_vcl_gtk3_a11y` that fails in a different way now,
but that's something for another day...)

Change-Id: Ic53dd8f87e3fa548a998363f5489819a02b2bf8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179862
Reviewed-by: Stephan Bergmann <stephan.bergmann@allotropia.de>
Tested-by: Jenkins