tdf#130857 qt weld: Introduce QtInstanceToolbar
Introduce QtInstanceToolbar as the native Qt implementation
for weld::Toolbar that uses a QToolBar widget.

This mostly adds the basic structure and lets Qt{,Instance}Builder
create an instance of the class when encountering a
"GtkToolbar" object in a .ui file.

The actual logic for the QtInstanceToolbar methods
still needs to be implemented. Calling them currently
triggers an assert.

Change-Id: I2734e5b931d5cb9608ec4d01222634a8ffada299
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179883
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>