cool#10630 lok doc sign: allow setting the pos of the Impress sign line
Open a PDF in LOK mode, insert a signature line, try to drag the
selected signature widget/shape, nothing happens.

Given that the inserted shape has a default position at page center, you
typically want to adjust that default position, but that requires adding
a few exceptions, since PDFs are normally read-only. Desktop Impress
gets around this since commit 5d296183072dc7cfe7a9985c38388b56f37d873c
(sd signature line: allow move / resize of shape before signing,
2020-06-26), but LOK clients want to rather dispatch an UNO command
(instead of handling this in the mouse event handler), so this requires
additional work.

A first problem is in isCommandAllowed() in desktop/ that explicitly
rejects the UNO command for read-only documents: allow this when a
signature widget is selected.  A second problem is in
SdrEditView::SetGeoAttrToMarked(), check for the signature widget there
as well.

Given that the implementation of .uno:TransformDialog already has a set
of flags to prevent doing anything for read-only documents, it looks
safe to allow the execution of the command itself.  The size of the
widget should be also possible to modify, that's not yet done here.

Change-Id: I92a60717e98ac33f7a1eb9770cd0c7903f60759c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180049
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins