tdf#164627 docx export: consolidate getWordCompatibilityMode()
Solves TODO: this is duplicated, better store it as DocxExport member?

It is generally useful to be able to query what compatibility mode
the DOCX will be exported as, so calculate the value once
and cache it.

I re-arranged things a bit, and I think it is clearer now
what is happening here.

Change-Id: I5cd619e1696ecef956a54957fb57f49803e7bc67
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179981
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>