tdf#41775 copy the Start Center menubar to the default menubar on macOS
The ShutdownIcon class on macOS handles the menubar for when there are
no windows displayed. So copy the entire menubar's tree to a native
menu and use that native menu as the default menubar.

Note: not all menu items may crash without a C++ window so currently
the Tools > Configure menu must force the Start Center window to
reappear before displaying its dialog.

Change-Id: I3190da0dffdd6c7181939aec71954e067689d081
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179980
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>