cool#10630 lok doc sign: allow setting the size of the Impress sign line
Insert a signature line in LOK mode when PEM certs are configured,
moving the just inserted shape works (even in the read-only PDF view),
but you can't resize the shape.

This is because moving goes via .uno:TransformDialog, while resizing
uses .uno:MoveShapeHandle.

Fix this similar to moving, by allowing the operation when the shape is
a just inserted signature line. Allowing the command in read-only mode
is meant to be safe, since the command implementation again checks if
the resize is allowed, and there we allow this only for signature lines.

Also expose the "is signature line" info on the LOK API, so it can show
the graphic selection handles conditionally (only when resize will be
allowed).

Change-Id: Ie85aa1211564758dc24ff83b4241de78a8e69ffc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180065
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins