tdf#152515 vcl: Implement Win32 vert CJK printing for all fonts
Previously, vertical CJK printing was implemented on Windows by
substituting the font with an @-prefixed alternate. This was a naming
convention once used by Microsoft to indicate that a font is a
specially-prepared vertical variant containing pre-rotated CJK glyphs.
However, modern fonts don't ship these variants and cannot be printed
correctly if this approach is used.

This change deletes the @-prefix substitution. Instead, a secondary
HFONT is used for printing upright characters, as determined by the
shaper. This implementation should work with all fonts, and the printed
output should now match the screen better.

Change-Id: I3c474980f90ade8475073926364ab59dd9310ca0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180014
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>
Tested-by: Jenkins