Related: tdf#163126 fix failure to restart print job
Commit 09484a1befd41681276cc9909b3228e5bf28a76e caused restarting
a print job to fail when either the "Print selection only" or
the "Brochure" checkboxes were changed.

Also, that same commit added logic to handle empty paper names
but setting the paper size would sometimes fail for standard
paper names so always set the paper size if the width and
height parameters are valid no matter what the paper name is.

Change-Id: Ia6bcbffca1152ba0c47309d0c208686d2d1b30e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180036
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins