math: store main viewshell id in math view shell
problem:
when we enter to edit a formula a new view shell is created.
In LOK all the JSwidgets would be mapped to this window id.
But when LOK requests widgets we get the requests from the main viewshell
and we can't find any widgets mapped to the main viewshell
which means LOK can't load any widgets.

Currently there's no mechanism to find the sub viewshell

Change-Id: I4f9ce161101ef6815f4d9e712d9bd42f0dc83ea3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180038
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180287
Tested-by: Jenkins