tdf#160051 PDF export: Artifact present inside tagged content
If a PDF file containing artifacts is added to a document as an image,
and the document is then exported as a tagged PDF, these artifacts are
placed into a structure element (e.g., figure), which is not allowed.
This fix removes unnecessary artifact tags from the content stream.

Change-Id: I590ebec9a7aecdaa42520008824469bc8a9ff65b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180041
Reviewed-by: Nagy Tibor <tibor.nagy.extern@allotropia.de>
Tested-by: Jenkins
(cherry picked from commit 05a075d23eb6003849e75582e12ef788e615a56d)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180157
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>