tdf#130857 qt weld: Support Calc "Move/Copy Sheet" dialog
Declare support for Calc's "Move/Copy Sheet" dialog that shows
up when right-clicking on a sheet in the bottom tabbar showing
sheets and selecting the "Move or Copy Sheet" context menu entry.

This means that native Qt widgets are used for that dialog
now when using the qt5 or qt6 VCL plugin and starting LO with
environment variable SAL_VCL_QT_USE_WELDED_WIDGETS=1 set.

Change-Id: Ibabfd1188a21c2d6b50595b8133ef297e995642d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180088
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>