tdf#164620 CRASH: selecting all and deleting
regression from
    commit 73a46895c5927d68a222b13dc811ea1cceb811a1
    Author: Noel Grandin <noel.grandin@collabora.co.uk>
    Date:   Thu Dec 5 07:58:44 2024 +0200
    tdf#119840 speed up SearchForStyleAnchor

We are dealing with a dangling pointer here, so I'm not sure if this
change will be sufficient.
We are very deep inside a series of listener callbacks and events here,
so no idea how to fix the actual problem i.e. not having a dangling
pointer at all.

Change-Id: I8314cd0f906e3b2652e10e55dfe8396abdf02856
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180111
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>