ScriptForge (SF_Calc new XRectangle property
The
   XReetangle(rangename)
property returns a com.sun.star.awt.Rectangle
structure describing the coordinates (in pixels)
on the screen where the given range is
located.

This opens the door to effective interactivity
in calc sheets with the use of popup menus.

Example:
   r = calc.XRectangle("B2")
   menu = CreateScriptService("PopupMenu", , _
           X := r.X + r.Width/2, Y := r.Y + r.Height/2)
initializes a popup menu centered on the given
cell.

The property exists in Basic and Python.
The SF_Calc help page should be updated
according to the actual patch.

Change-Id: I2a97c5c7b5bf84d36a2aab5e4b822e30c9d53d8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180140
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins