tdf#130857 qt weld: Convert accelerator in button label
Otherwise e.g. the "~Open" label (STR_OPEN, which
gets set in SfxTemplateManagerDlg::SfxTemplateManagerDlg)
gets shown as a literal "~Open", rather than "Open" with
the "O" being the accelerator key in a WIP branch experimenting
with Qt welding of the "File" -> "Templates" -> "Manage Templates"
dialog.

While at it, simplify the method a bit to just
run the single method in the main thread
unconditionally.

Change-Id: Iea25948f77edb6e884bad96cb36ab7eec67a141e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180253
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins