tdf#34837 sc: Use border from correct cell for merged cells in RTL
This change fixes a bug causing borders to fail to render for merged
cells in RTL spreadsheets.

The root cause for this bug was framelinkarray utility code using the
top-left cell as the source for border styles of merged cells. This use
was correct for LTR spreadsheets. However, this framelinkarray data is
mirrored for RTL documents. After mirroring, the correct cell containing
border styles is the top-right one.

This change also reverts and reimplements a prior fix for tdf#135843
(commit 586a0f149f332c0b0e53c0bb30568d4bd411b0e3), which violated the
framelinkarray contract that edge styles for merged cells must be added
to the top-left underlying cell of a merged cell.

Change-Id: I27eec416d54f9f99cd5df1151a12c758f350c789
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180256
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>