tdf#164500 docx import framePr: add blank para as anchor before table
In my unit test there is a "problem"
because the added paragraph wraps below the frame
and thus pushes the content down an extra line (compared to MS Word).
But if the frame was not autosized (full width),
then the added paragraph would wrap around behind it perfectly.
This is emulation (since we can't anchor to a table start node),
so it will have to be good enough.

make CppunitTest_sw_ooxmlexport18 \
    CPPUNIT_TEST_NAME=testTdf164500_framePrBeforeTable

Change-Id: I1318dbacf853fb7d08c988e48dda7ac20c9498ab
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/179492
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins