cool#10630 doc electronic sign: fix no graphic selection for the signature line
Insert a signature line in "extern" mode, the shape gets selected but
there is no graphic selection at a LOK API level.

This is because GetSignPDFCertificate() returned an XCertificate, which
is empty in the external signing case, so we can't differentiate between
no signing and external signing.

Fix this by changing the return type to svl::crypto::CertificateOrName,
this way SdrMarkView::SetMarkHandlesForLOKit() can annotate the
signature line correctly even in the external signing case.

The tracking of the signature line selection is still in the model (not
in the view), that's not yet fixed here.

Change-Id: I4ef9c1fa0a88af0c0fcd55156b973a3705f985c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180296
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins