tdf#164622 - SVG: fix broken cropped svg files at metafile export
For creating metafile, there mask polygon from MaskPrimitive2D needs to be closed.
Otherwise later during the metafile export the we will skipp the rendering of the
last points of the polylines.

SVG clipping paths: "A clipping path is a set of closed vector path that may consist
of an arbitrary number of straight and curved segments."

Change-Id: I21487a81b904ec7ec5c6bd2a89b7539aeb86d606
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180278
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>
Tested-by: Jenkins