tdf#164141 Fix command button xls to xlsx convert problem.
MSO has two type controls. ActiveX and normal. ControlTypeinMSO
specifies the form control object is ActiveX or not.

2: ActiveX form control
1: Normal form control

Current case is simple ActiveX button created by MSO can not be exported by
LibreOffice properly. It was triggering the repair dialog on MSO. And we
even lost the button in LibreOffice.

If the pCurrXclObj is failed, we need to give a chance to export button
by using XclExpTbxControlObj instead of directly XclObjAny

This patch just fixes the Command button xls to xlsx convertion problem.
Not all the form controls.

Signed-off-by: Gülşah Köse <gulsah.kose@collabora.com>
Change-Id: Ifabf4d0e16bd620f8201a2143d8af7c64d8859a6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180414
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/181345