Use IID_PPV_ARGS and CComPtr to simplify
IID_PPV_ARGS allows to avoid reinterpret_casts (and also explicit
auto-deducible IIDs). Also, there always was a mismatch between the
type of mWebBrowser2 (which is IWebBrowser2) and the queried IID in
CSOActiveX::SetClientSite (which was IID_IWebBrowser, instead of
the correct IID_IWebBrowser2). Looks like in practice it wasn't a
problem, but now it's consistent.

Use of CComPtr in CSOActiveX::Load allows to release the pointer
correctly (previously, the call to Release was missing).

Change-Id: I29d672052780a36465f6143106c22fa2c90dab5a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180448
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>