tdf#130857 qt weld: Set GtkButton props for MenuButton
The GTK 3 GtkMenuButton [1] subclasses GtkButton [2].
Therefore, let QtBuilder apply the GtkButton properties for the
QToolButton created for a "GtkMenuButton" object in a .ui file as well.

This e.g. ensures that the menu button in the
"File" -> "Templates" -> "Manage Templates" dialog
gets the "Manage" label in a WIP branch where support
for that dialog using native Qt widgets is declared.

[1] https://docs.gtk.org/gtk3/class.MenuButton.html
[2] https://docs.gtk.org/gtk3/class.Button.html

Change-Id: I3f2e7b701776476a6677d2bb324ffbdd4956f471
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180461
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins