tdf#151423 use the same modifier keys during a scrolling session
Revert commit f5ef5eafdf70a36edd5129147502a9c74df89456 as it
completely disabled the ability to zoom on mice with limited
support for gestures such as the Apple Magic Mouse by pressing
pressing the Command key while scrolling. So try a different
approach and use the modifier keys that were pressed when the
current scrolling session was started for the entire scrolling
session.

Unfortunately, session state changes such as a scrolling session
began or ended are not fired when scrolling with a regular mouse
scrollwheel so use a significant pause between scrolling events
as a rough indicator that a new scrolling session has begun.

Change-Id: I898d7f0d3e28ad0436fcb75504f9efb6bf9cdb7d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180451
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>