tdf#130857 templatedlg.ui: Don't use "pack-type end"
Don't use the "pack-type" property with a value of "end"
to reverse the order of GtkBox children, but use the
default (of "start") and set the indices for the "packing"
properties of the children in the usual order, which is more
straighforward.

The "pack-type" GtkBox property no longer exists in GTK 4 (see [1];
`Convert3To4` in vcl/unx/gtk4/convert3to4.cxx currently
has special handling for it for the "GtkBox" case to revert
the order) and I also think it makes more sense to "normalize"
the .ui files than implementing such a special handling in
QtBuilder.

Adjusting this in glade and resaving the .ui file also results
in the GtkBox children to be defined in visual order, which ensures
that the (focus) order is correct in Qt as it (by default) is based
on the order in which widgets are created.

(The "File" -> "Templates" -> "Manage Templates" dialog is not
yet supported by QtInstanceBuilder, just noticed this while
experimenting in a WIP branch.)

[1] https://discourse.gnome.org/t/gtk4-replacement-for-pack-type/5574

Change-Id: I1633fc05da68a7ccce36e17d13d0d654f4e6145b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180463
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>