tdf#130857 qt weld: Set GtkButton props for ToggleButton
The GTK 3 GtkToggleButton [1] subclasses GtkButton [2].
Therefore, let QtBuilder apply the GtkButton properties for the
QToolButton created for a "GtkToggleButton" object in a .ui file
as well.

This e.g. ensures that the icons for the toggle buttons in the
"File" -> "Templates" -> "Manage Templates" dialog are set
in a WIP branch where support for that dialog using native Qt
widgets is declared.

[1] https://docs.gtk.org/gtk3/class.ToggleButton.html
[2] https://docs.gtk.org/gtk3/class.Button.html

Change-Id: I1efedea116a799b8bc41fa81a27585af65893220
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180465
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>