tdf#130857 qt weld: Support QtInstanceToggleButton::set_inconsistent(false)
While QCheckBox supports an "inconsistent" state
(see QCheckBox::tristate), I don't see that for
QToolButton/QAbstractButton.

Therefore, keep triggering an assert when trying to
enable the indeterminate state (and reconsider what to
do whenever encountering that example while working
on support for more dialogs in the future).

However, explicitly *disabling* the inconsistent
state is fine: Don't do anything in that case.

This gets called by the subclass, weld::Toggleable::set_state
when setting the toggle button to checked or unchecked.

[1] https://doc.qt.io/qt-6/qcheckbox.html#tristate-prop

Change-Id: I4cb1475a4b5e3295ada7ed18dcbfcbf24baa53e9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180466
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>