ScriptForge (SF_Calc new Intersect() method
The
    calc.Intersect(range1, range2) method
computes the common area as a string
which is a sub-range of both input ranges.
It returns an empty string when the
sub-range is empty.

Potential use case: customizing a popup
or context menu depending on the presence
of the selected cell in a given area.

In addition, the A1Style() method returns
the sheetname in the output string only
when it is an explicit argument.

Changes are applicable both for Basic and
Python scripts.

They require an update of the user documentation.

Change-Id: I920fc8bee6dace9ceed263a0570b4c0197bc8cb8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180479
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins