cool#10630 doc electronic sign: unselect & reject reselect
Insert a signature line, move/resize the selected shape, finish signing,
the signature line is still selected and you can still move/resize it.

This was working for the desktop visual signing which
SfxViewShell::GetSignPDFCertificate() returned an XCertificate, but not
for electronic signing, which is a separate branch in
SfxObjectShell::ExecFile_Impl().

Fix the problem by deselecting the shape even in the desktop case and
the moving this cleanup code to a new ResetSignatureSelection() and
calling that in the electronic sign case, too.

SdrMarkView::UnmarkAllObj() is private to sfx2/, so go via UNO to do the
same.

Change-Id: I52c89c4aea0203f8624bd801a900186211541dc3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180508
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>